Fix: avoid int overflow with very large size arrays

+++ C/src/TEOBResumSUtils.c
@@ -948,7 +948,7 @@ void set_multipolar_idx_mask (int *kmask, int n, const int *idx, int m, int on)
 /** Compute size of a uniform grid t0:dt:tf */
 int get_uniform_size(const double tN, const double t0, const double dt)
 {
-  return ((int)((tN - t0)/dt + 1)); 
+  return ((long int)((tN - t0)/dt + 1)); 
 }
 
 /** Intersection of two int arrays */
